Refactor sequential reads to count blocks. Relying on a read to return no bytes was a fragile way to do wrap-around.
Sequential read wrap-around was depending the EOF behavior of read() to
return 0 bytes.  That isn't the most robust approach possible.  As one
example, in cases cases where files/devices were not an exact multiple
of block_size, this was counting the leftover small reads at the end
of the file/device as a full read.  That could inflate results a good
bit at the end of runs where the block size became very large.

Counting blocks also allows the seq read logic to be refactored to
start at any position.  The way all the sequential read threads were
synchronized to start at the same spot is itself a problem.  This
fix was needed first.

New code was tested with some temporary logging and a test long enough
here to wrap around to the start:

dd if=/dev/zero of=test.bin bs=1M count=1000
./iops -n 1 -b 512 -p sequential -t 20 test.bin

1468848531.4 count=2047997 seq to pos 2047997 out of 2048000 blocks
1468848531.4 count=2047998 seq to pos 2047998 out of 2048000 blocks
1468848531.4 count=2047999 seq to pos 2047999 out of 2048000 blocks
1468848531.4 count=2048000 seq to pos 0 out of 2048000 blocks
1468848531.4 count=2048001 seq to pos 1 out of 2048000 blocks
1468848531.4 count=2048002 seq to pos 2 out of 2048000 blocks

Here's the temporary debug lines:

          if pattern=='random':
              pos = random.randint(0, mediasize - blocksize) # need at least one block left
              pos &= ~(sectorsize-1)   # sector alignment at blocksize
              fh.seek(pos)
+         else:
+             print time.time(), "count=%d seq to pos %d out of %d blocks" % (count,seq_block,blocks)